Move variant callers to analysis api
