+++ cg/meta/workflow/balsamic.py
@@ -18,6 +18,7 @@ from cg.io.controller import ReadFile
 from cg.meta.workflow.analysis import AnalysisAPI
 from cg.meta.workflow.fastq import BalsamicFastqHandler
 from cg.models.balsamic.analysis import BalsamicAnalysis
+from cg.models.balsamic.config import BalsamicVarCaller
 from cg.models.balsamic.metrics import (
     BalsamicMetricsBase,
     BalsamicTargetedQCMetrics,
@@ -630,3 +631,34 @@ class BalsamicAnalysisAPI(AnalysisAPI):
     def get_genome_build(self, analysis_metadata: BalsamicAnalysis) -> str:
         """Returns the reference genome build version of a Balsamic analysis."""
         return analysis_metadata.config.reference.reference_genome_version
+
+    @staticmethod
+    def get_variant_caller_version(var_caller_name: str, var_caller_versions: dict) -> str | None:
+        """Return the version of a specific Balsamic bioinformatic tool."""
+        for tool_name, versions in var_caller_versions.items():
+            if tool_name in var_caller_name:
+                return versions[0]
+        return None
+
+    def get_variant_callers(self, analysis_metadata: BalsamicAnalysis) -> list[str]:
+        """
+        Return list of Balsamic variant-calling filters and their versions (if available) from the
+        config.json file.
+        """
+        sequencing_type: str = analysis_metadata.config.analysis.sequencing_type
+        analysis_type: str = analysis_metadata.config.analysis.analysis_type
+        var_callers: dict[str, BalsamicVarCaller] = analysis_metadata.config.vcf
+        tool_versions: dict[str, list] = analysis_metadata.config.bioinfo_tools_version
+        analysis_var_callers = list()
+        for var_caller_name, var_caller_attributes in var_callers.items():
+            if (
+                sequencing_type in var_caller_attributes.sequencing_type
+                and analysis_type in var_caller_attributes.analysis_type
+            ):
+                version: str = self.get_variant_caller_version(
+                    var_caller_name=var_caller_name, var_caller_versions=tool_versions
+                )
+                analysis_var_callers.append(
+                    f"{var_caller_name} (v{version})" if version else var_caller_name
+                )
+        return analysis_var_callers
