+++ tests/meta/report/test_balsamic_api.py
@@ -50,51 +50,6 @@ def test_get_sample_metadata(
     assert sample_metadata.model_dump() == expected_metadata
 
 
-def test_get_variant_callers(report_api_balsamic, case_id):
-    """Test variant callers extraction for TN-PANEL analysis."""
-
-    # GIVEN a mock metadata object
-    balsamic_metadata = report_api_balsamic.analysis_api.get_latest_metadata(case_id)
-
-    # GIVEN an expected output for tumor normal panel analysis
-    expected_callers = [
-        "manta (v1.6.0)",
-        "cnvkit",
-        "vardict (v2019.06.04=pl526_0)",
-        "dnascope",
-        "tnhaplotyper",
-        "manta_germline (v1.6.0)",
-        "haplotypecaller",
-        "TNscope_umi",
-        "delly (v0.8.7)",
-    ]
-
-    # WHEN retrieving the analysis variant callers
-    variant_callers = report_api_balsamic.get_variant_callers(balsamic_metadata)
-
-    # THEN check that the callers are correctly identified
-    assert variant_callers == expected_callers
-
-
-def test_get_variant_caller_version(report_api_balsamic, case_id):
-    """Tests variant caller version extraction."""
-
-    # GIVEN a tool name and a mock variant caller versions dictionary
-    var_caller_name = "manta"
-    var_caller_versions = report_api_balsamic.analysis_api.get_latest_metadata(
-        case_id
-    ).config.bioinfo_tools_version
-
-    # GIVEN the tools mock version
-    expected_version = "1.6.0"
-
-    # WHEN retrieving the version of a specific variant caller
-    version = report_api_balsamic.get_variant_caller_version(var_caller_name, var_caller_versions)
-
-    # THEN verify that the extracted version is correct
-    assert version == expected_version
-
-
 def test_is_report_accredited(report_api_balsamic, case_id):
     """Test report accreditation for a specific BALSAMIC analysis."""
