+++ tests/meta/workflow/conftest.py
@@ -21,8 +21,15 @@ from tests.cli.workflow.balsamic.conftest import (
     fastq_file_l_2_r_2,
 )
 from tests.meta.compress.conftest import compress_api, real_crunchy_api
+from tests.mocks.balsamic_analysis_mock import MockBalsamicAnalysis
 from tests.mocks.tb_mock import MockTB
 from tests.store_helpers import StoreHelpers
+
+
+@pytest.fixture(scope="function")
+def analysis_api_balsamic(cg_context: CGConfig) -> MockBalsamicAnalysis:
+    """BALSAMIC ReportAPI fixture."""
+    return MockBalsamicAnalysis(cg_context)
 
 
 @pytest.fixture(scope="function")
